Clean up the bevy example a little
Move the web asset handling into a separate module, so that main focuses on bevy.
